requirements.txt
PyQt5
opencv-python
numpy
pyautogui
Pillow
pytesseract
